separate structs out into files
